Remove the pointless DynamicMethod ...Cast methods
They are pointless because the generic cast will never cause a
ClassCastException and therefore the methods give no advantage over the
other two. Especially the generic cast would only be useful to assign
the returned value to the correct type, but then this type could also be
used as class of the method return type and therefore the generic type T
of the DynamicMethod object itself.

If your code used one of the methods, replace those uses with the
regular methods. They are better in every way.

In addition to that, the ...Cast methods caught a ClassCastException
even if it occurred in the invoked method.