Add strictOrder for parameters in structural tests (#156)
The structural tests are often used to check whether the student implemented the expected structure and especially methods. This simplifies the behavior tests, since reflections can be avoided. Currently, the order of parameters is irrelevant for checking methods. This grants the students more freedom, but can make the behavior tests more complicated. This PR adds the "strictOrder" tag to methods and constructors, which forces the student to exactly follow the expected order of parameters.