Introduce new TrustedThreads annotation to specify thread whitelisting
For jqwik, an extended thread trust scope is needed as jqwik uses the
common pool for property execution since version 1.3.5. However, Ares
does not trust the common pool. To keep compatibility and security, we
introduce a new annotation to specify which threads Ares can trust. This
allows the usage of new jqwik versions, such as 1.6.0.