Try to initialize all common pool workers in the main thread group
If they happen to be in the test thread group, Ares will try to
terminate the treads, wich can lead to a non-functioning common pool. We
can consider adding that back as an option, but usually destroying the
common pool is not a good option.

Here we simply use a big enough parallel int stream to spawn all worker
threads.