Check and interrupt the common pool better to prevent malfunction
Common pool workers have a fixed name, and therefore, we can identify
them based on their names. The old approach using the thread groups was
faulty because the thread group depends on whether the common pool was
started with SecruityManager. Since this differs for IDEs and Build
Tools or even test execution, we cannot rely on the thread group and
just use the name instead.
In addition to that, we switch to interrupting only common pool threads
that are executing non-whitelisted code to avoid interrupting common
pool workers that are already idle.