Make all major Ares annotations inherited, this resolves #124
How this works now:

We first search at the test method, then at the containing class, all
its superclasses (but not interfaces!) the hierarchy up, and then in
case of nested classes the next class outwards (and then again all
superclasses of that class) and so on until the outermost class is
reached. The upwards class hierarchy traversal is the new component
here.

This means:

For non-repeatable annotations (such as Deadline), we stop as soon as we
find the first Deadline using this process. For repeatable annotations
(such as WhitelistPath), we collect all annotations we can find using
this process.

The PublicTest and HiddenTest annotations are unchanged because they are
meant for methods only (and we generally recommend using just Public and
Hidden).